+++ src/i8c/constants.py
@@ -37,7 +37,6 @@ I8_TYPE_OPAQUE = "o"
 I8_TYPE_FUNC = "F"
 
 I8_EXT_FUNCTION = "f"
-I8_EXT_SYMBOL = "s"
 
 DW_OP_addr = 0x03
 DW_OP_deref = 0x06